requirements.txt
colorlog==6.9.0
homeassistant==2025.2.5
pip>=21.3.1
ruff==0.11.6
aiohttp
voluptuous
paho.mqtt